+++ src/controllers/index.ts
@@ -1,8 +1,10 @@
 import userController from './user.controller';
+import authController from './auth.controller';
 
-export { userController };
+export { userController, authController };
 
 // Export for easier imports
 export default {
   userController,
+  authController,
 };