+++ src/validations/index.ts
@@ -1,7 +1,9 @@
 import * as userValidation from './user.validation';
+import * as authValidation from './auth.validation';
 
-export { userValidation };
+export { userValidation, authValidation };
 
 export default {
   userValidation,
+  authValidation,
 };